\name{libaglm-package}
\alias{libaglm-package}
\alias{libaglm}
\docType{package}
\title{
\packageTitle{libaglm}
}
\description{
\packageDescription{libaglm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{libaglm}
\packageIndices{libaglm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{libaglm}

Maintainer: \packageMaintainer{libaglm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
